Bug fixes and new commands
s!ctgp has been updated to provide more resources such as links to the patchers and the downloads for legacy versions.
s!wifi has been added, able to provide resources for Wi-Fi rankings.
A bug that flagged the Players' Page database as in maintenance while it was not has been fixed.